requirements.txt

tabulate
streamlit
python-dotenv
langchain-groq
langchain-experimental
pandas
plotly
